build: Add the platform tags to container builds
This adds the platform tags via the Makefile rather than the CI file, to
better let us build it in non-CI environments.

The previous setup was because of incompatibilities between
podman/docker and the weird behaviour of the --platform agruments. It
would previously attempt to pull the "wrong" image and fail (wrong
architecture) and thus needed to be told "pull-never" and then have the
override of pre-pulling the container.

Now, with current podman, things seem to work again without ugly
work-arounds.