Migrate to using our standard "container" build
As we now have a functional arm32 CI pipeline in gitlab with
qemu-user-static, we can fall back to our normal container image.

There is a small hack in using "PODMAN_PULL" to inject the target
architectures into the system when building the container, which is
necessary to make the intermediate cache layers use the right containers
and not fail to resolve.