Restore libdbus to build images
Due to an unexpected change when removing --no-install-recommends on
debian9, the expected libdbus library isnt part of the debian9.

This was implicitly used by the modio-networks use of the debian9 image
that is built to test taht we have the right executable environment.

This restores the package, and adds it to all other build images to
preserve consistency.